pyproject.toml
[tool.poetry]
name = "roborabbit"
version = "0.2.1"
description = "Set up your rabbit instance using a declarative yaml file."
authors = ["Skyler Lewis <skyler@hivewire.co>"]
license = "Apache-2.0"
readme = "README.md"
homepage = "https://github.com/alairock/roborabbit"
repository = "https://github.com/alairock/roborabbit"
documentation = "https://github.com/alairock/roborabbit/blob/main/README.md"

[tool.poetry.dependencies]
python = "^3.9"
click = "^8.0.1"
PyYAML = "^5.4.1"
aio-pika = "^6.8.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.scripts]
roborabbit = "roborabbit.main:main"
